Movi sa parti in python aìnturu de sa cartelledda cosa sua.
Former-commit-id: 69d4388048ef8f1e14806c47a7ff1757b07104cb
