+++ python/environment.yml
@@ -1,0 +1,13 @@
+name: data-science
+channels:
+  - conda-forge
+dependencies:
+  - python=3.8
+  - matplotlib=3.3.1=0
+  - numpy>=1.19.1
+  - scipy=1.5.2
+  - pandas>=1.1.1
+  - plotly>=4.10.0
+  - jupyter
+  - geopandas
+  - descartes
